+++ packages/ubq/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - mail-queue.queueMail method: remove unnecessary `fieldList: ['ID']` in queue insertion 
+  to prevent call of selectAfterInsert method since is ID not used anywhere 
 
 ### Deprecated