+++ packages/ubq/CHANGELOG.md
@@ -8,14 +8,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
- - mail-queue.queueMail method: remove unnecessary `fieldList: ['ID']` in queue insertion 
-  to prevent call of selectAfterInsert method since ID is not used 
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+## [5.3.3] - 2020-03-17
+### Changed
+ - mail-queue.queueMail method: remove unnecessary `fieldList: ['ID']` in queue insertion 
+  to prevent call of selectAfterInsert method since ID is not used 
 
 ## [5.3.2] - 2020-03-09
 ## [5.3.1] - 2020-03-04